Fix consistency recovery.
For some transactions we only have to rewrite the response (i.e. the request is present before the consistent timestamp). Previously these requests were truncated and not rewritten. Now they are not truncated anymore.

Also ensure that the first rewritten transaction is the one of the last consistent timestamp. Before the rewritten transactions were ordered by timestamp. If the first transaction to rewrite had a smaller timestamp, the log wasn't truncated at the right position.