Service member must go down on consistency error in order to recover
Even if transaction log is not enabled, the service member must go down on consistency error in order to recover (i.e. to go up again). The TransactionRecorder prevent further message to be appended forever. The failed recorder is destroyed when the service member goes down and recreated when going up again.