Expose writen consistent timestamp instead of computed value
Transaction log replication source must not read beyong the current member consistent timestamp to prevent reaching the end of the log file. This is possible that no log record has been written yet with the computed consistent timestamp and thus allowing the source to reach the end of the log file. TransactionRecorder now exposes the the last written consistent timestamp to prevent that.